src/Game/reducer.test.ts
import { reduceGameState, createGameState } from './reducer';
import { GameActionType } from './actions';
import { GameStatus, GameState } from './types';

describe('reduceGameState', () => {
  let initialState: GameState;

  beforeEach(() => {
    initialState = createGameState({ /* your game config here */ });
  });

  it('should handle START_COUNTDOWN action', () => {
    initialState.status = GameStatus.READY;
    const action = { type: GameActionType.START_COUNTDOWN };
    const newState = reduceGameState(initialState, action);
    expect(newState.status).toBe(GameStatus.COUNTDOWN);
  });

  it('should not change state for START_COUNTDOWN if status is not READY', () => {
    initialState.status = GameStatus.LOADING;
    const action = { type: GameActionType.START_COUNTDOWN };
    const newState = reduceGameState(initialState, action);
    expect(newState).toBe(initialState);
  });

  it('should handle START_PLAYING action', () => {
    initialState.status = GameStatus.COUNTDOWN;
    const action = { type: GameActionType.START_PLAYING };
    const newState = reduceGameState(initialState, action);
    expect(newState.status).toBe(GameStatus.PLAYING);
    expect(newState.startTime).toBeDefined();
  });

  it('should not change state for START_PLAYING if status is not COUNTDOWN', () => {
    initialState.status = GameStatus.READY;
    const action = { type: GameActionType.START_PLAYING };
    const newState = reduceGameState(initialState, action);
    expect(newState).toBe(initialState);
  });

  it('should handle PROCESS_INPUT action', () => {
    initialState.status = GameStatus.PLAYING;
    const action = { type: GameActionType.PROCESS_INPUT, payload: { /* input data */ } };
    const newState = reduceGameState(initialState, action);
    // Add assertions based on how input is processed
    expect(newState).toBe(initialState); // Assuming no state change for now
  });

  it('should not change state for PROCESS_INPUT if status is not PLAYING', () => {
    initialState.status = GameStatus.READY;
    const action = { type: GameActionType.PROCESS_INPUT, payload: { /* input data */ } };
    const newState = reduceGameState(initialState, action);
    expect(newState).toBe(initialState);
  });

  it('should handle SHOW_RESULTS action', () => {
    initialState.status = GameStatus.PLAYING;
    const action = { type: GameActionType.SHOW_RESULTS };
    const newState = reduceGameState(initialState, action);
    expect(newState.status).toBe(GameStatus.RESULTS);
  });

  it('should not change state for SHOW_RESULTS if status is not PLAYING', () => {
    initialState.status = GameStatus.READY;
    const action = { type: GameActionType.SHOW_RESULTS };
    const newState = reduceGameState(initialState, action);
    expect(newState).toBe(initialState);
  });
});